refactor/AB#82357_create-shared-error-handler-logic-for-the-app feat: handle graphQL errors using apollo links
