+++ libs/shared/src/lib/interceptors/error-handler/error-handler-interceptor.service.ts
@@ -4,11 +4,9 @@ import {
   HttpHandler,
   HttpEvent,
   HttpInterceptor,
-  HttpResponseBase,
 } from '@angular/common/http';
-import { Observable, catchError, of, switchMap, throwError } from 'rxjs';
+import { Observable, catchError, throwError } from 'rxjs';
 import { SnackbarService } from '@oort-front/ui';
-import { isNil } from 'lodash';
 
 /**
  * Error handler interceptor service
@@ -39,27 +37,6 @@ export class ErrorHandlerInterceptorService implements HttpInterceptor {
   ): Observable<HttpEvent<unknown>> {
     if (request.url.includes(this.environment.apiUrl)) {
       return next.handle(request).pipe(
-        // GraphQl errors would be treated as so for the app even if status is 200
-        switchMap((req) => {
-          const keysFromBody = Object.keys((req as any).body?.data);
-          if (
-            req instanceof HttpResponseBase &&
-            // Could be cases where we query two different objects and one can fail and the other no
-            // In those cases we prioritize the use of correctly retrieved data
-            // Therefor we treat as strict error if all queries fail
-            (((req as any).body?.errors?.length &&
-              keysFromBody?.every((key) =>
-                isNil((req as any).body?.data[key])
-              )) ||
-              keysFromBody?.every((key) => isNil((req as any).body?.data[key])))
-          ) {
-            return throwError(() => {
-              return { error: { errors: (req as any).body?.errors } };
-            });
-          } else {
-            return of(req);
-          }
-        }),
         catchError((res: any) => {
           // Open default snackbar if error is not a GraphQL type
           if (res.status) {
@@ -76,11 +53,7 @@ export class ErrorHandlerInterceptorService implements HttpInterceptor {
               duration: 0,
             });
           }
-          return throwError(() => {
-            return !isNil(res.message)
-              ? { errors: [res.message] }
-              : res.error.errors;
-          });
+          return throwError(() => res.error.errors);
         })
       );
     } else {
