+++ libs/shared/src/lib/services/form-builder/form-builder.service.ts
@@ -16,7 +16,7 @@ import { BehaviorSubject } from 'rxjs';
 import { SnackbarService } from '@oort-front/ui';
 import { FormHelpersService } from '../form-helper/form-helper.service';
 import { HttpClient } from '@angular/common/http';
-import { errorMessageFormatter } from '../../utils/graphql/error-message-formatter';
+import { errorMessageFormatter } from '../../utils/graphql/error-handler';
 
 /**
  * Shared form builder service.
