+++ libs/shared/src/lib/services/form-helper/form-helper.service.ts
@@ -22,7 +22,7 @@ import {
   DELETE_DRAFT_RECORD,
   EDIT_DRAFT_RECORD,
 } from './graphql/mutations';
-import { errorMessageFormatter } from '../../utils/graphql/error-message-formatter';
+import { errorMessageFormatter } from '../../utils/graphql/error-handler';
 /**
  * Shared survey helper service.
  */
