--- libs/shared/src/lib/utils/graphql/error-message-formatter.ts
@@ -1,8 +1,0 @@
-/**
- * Gets the message in the graphql error
- *
- * @param errors GraphQL error
- * @returns formatted message
- */
-export const errorMessageFormatter = (errors: any): string =>
-  (errors.networkError || errors)?.[0].message ?? '';
