+++ libs/shared/src/lib/utils/public-api.ts
@@ -1,4 +1,4 @@
 export * from './validators/cron.validator';
 export * from './graphql/connection.type';
-export * from './graphql/error-message-formatter';
+export * from './graphql/error-handler';
 export * from './update-queries';
